+++ js/tts.js
@@ -16,7 +16,16 @@ function sendMsgToCS(tabNumber, message) {
 
 $(document).ready(async function () {
 
-    // alert($("#scrollSwitch").attr('checked'));
+    // load switch value
+    var scrollSwitchVal = await getVal("scrollSwitch");
+    $('#scrollSwitch').prop('checked', scrollSwitchVal);
+
+    // save switch value when switch is changed
+    $('#scrollSwitch').change(function() {
+        chrome.storage.sync.set({ scrollSwitch: $(this).prop('checked') })
+        var msg = 'scroll changed to ' + JSON.stringify($(this).prop('checked'));
+        sendMsgToCS(0, msg);
+    })
 
     // set the currentPage to defaultPopup when the back button is clicked
     $('#backLink').click(function () {
@@ -56,7 +65,6 @@ $(document).ready(async function () {
             option.setAttribute('data-lang', voices[i].lang);
             option.setAttribute('data-name', voices[i].name);
             voiceSelect.appendChild(option);
-            console.log(voices[i].localService)
         }
     }