+++ changelog.md
@@ -6,6 +6,7 @@ All changes made after 05/30/2020 will be documented in this file.
  - code to add hyphenated words together to be used in tts (because they are split by hyphen by splitting.js)
  - auto scrolling switch html/css
  - universal function to get values from chrome storage
+ - scrolling switch controls the autoscroll(successfully)
 
 ## [06/16/2020]
 ### Added